Refactor: Move common state parts to state.go
This is just shuffling pieces of code around to make the next commit easier to grasp.
There were close to zero changes, primarily some lines moved to a new file.
